Fitness = Tuple + Counter (WIP)
Fitness can now be defined to consist of an arbitrary tuple of primitive types.
The Counter is appended in CandidateBase. The first element is considered
principal and is being compared by operator<().

This has the advantage that we don't need to care about e.g. controls when
we're not interested.

New interface is work in progress and is currently only implemented in Simple,
other problems will fail.